+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 2.3.6, 6/16/2022
+
+### Added
+
+1. Support multiple pub/sub cluster instances
+2. Added cloud connection life cycle service
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 2.3.5, 5/30/2022
 
 ### Added